// api/src/common/enums/id_type.enum.ts
export enum IdType {
  CITIZENSHIP_CARD = 'Cédula de Ciudadanía',
  FOREIGNER_ID = 'Cédula de Extranjería',
  IDENTITY_CARD = 'Tarjeta de identidad',
  CIVIL_REGISTRATION = 'Registro Civil',
  PASSPORT = 'Pasaporte',
  SPECIAL_RESIDENCE_PERMIT = 'Permiso Especial de Permanencia',
  TEMPORARY_PROTECTION_PERMIT = 'Permiso por Protección Temporal',
}